Remove unused EnvLoader files and update UltrasonicModule to return float values; add Google Sheet task and LED task implementations.

+++ src/UltrasonicModule.h
@@ -19,8 +19,8 @@
 void initUltrasonicSensors();
 
 // Functions to get distances from each Ultrasonic sensor
-int getUltrasonic1Value();
-int getUltrasonic2Value();
-int getUltrasonic3Value();
+float getUltrasonic1Value();
+float getUltrasonic2Value();
+float getUltrasonic3Value();
 
 #endif // ULTRASONIC_MODULE_H
